Fix merge so that tests don't regress
